fix: rollback to 2.0.9

+++ src/lcd/extui/mks_ui/draw_wifi_tips.h
@@ -25,6 +25,7 @@
   extern "C" {
 #endif
 
+
 void lv_draw_wifi_tips();
 void lv_clear_wifi_tips();
 
@@ -37,7 +38,7 @@ extern TIPS_TYPE wifi_tips_type;
 
 typedef struct {
   unsigned char timer;
-  uint16_t timer_count;
+  unsigned int timer_count;
 } TIPS_DISP;
 extern TIPS_DISP tips_disp;
